Reorganizing and Verse References
The code for cleaning the texts has been moved to its own file to be less confusing. This includes splitting into verses along with sentences now.

Each of the files can now be easily switched between books, as the top of the main function has a variable for which book or books the file will use.